% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-shinytest2.R
\name{app_driver_expect_picks_visible}
\alias{app_driver_expect_picks_visible}
\alias{app_driver_expect_picks_hidden}
\title{Expect that a CSS selector resolves to at least one visible element.}
\usage{
app_driver_expect_picks_visible(selector, app_driver, timeout)

app_driver_expect_picks_hidden(selector, app_driver, timeout)
}
\arguments{
\item{selector}{\code{character(1)} CSS selector to check.}

\item{app_driver}{App driver object.}

\item{timeout}{\code{numeric(1)} maximum wait time.}
}
\value{
Called expectation result.
}
\description{
Use this helper only for tests and development workflows.
}
\section{Functions}{
\itemize{
\item \code{app_driver_expect_picks_hidden()}: Check if a selector is hidden for a given timeout.

}}
\examples{
\dontrun{
if (requireNamespace("shinytest2", quietly = TRUE)) {
  data <- within(teal.data::teal_data(), iris <- iris)
  test_picks <- picks(
    datasets("iris"),
    variables(choices = c("Sepal.Length", "Sepal.Width"), selected = "Sepal.Length")
  )
  teal_app <- teal::init(
    data = data,
    modules = teal::modules(tm_merge(label = "badge test", picks = list(pick = test_picks)))
  )
  app_driver <- suppressWarnings(shinytest2::AppDriver$new(
    shiny::shinyApp(ui = teal_app$ui, server = teal_app$server)
  ))
  app_driver$wait_for_idle()
  app_driver_expect_picks_hidden("[id$='inputs_container']", app_driver, timeout = 1000)
  app_driver$click(selector = "[id$='inputs-summary_badge']")
  app_driver_expect_picks_visible("[id$='inputs_container']", app_driver, timeout = 1000)
  app_driver$stop()
}
}

}
