% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-shinytest2.R
\name{app_driver_expect_visible}
\alias{app_driver_expect_visible}
\alias{app_driver_expect_hidden}
\title{Expect that a CSS selector resolves to at least one visible element.}
\usage{
app_driver_expect_visible(selector, app_driver, timeout)

app_driver_expect_hidden(selector, app_driver, timeout)
}
\arguments{
\item{selector}{\code{character(1)} CSS selector to check.}

\item{app_driver}{App driver object.}

\item{timeout}{\code{numeric(1)} maximum wait time.}
}
\value{
Called expectation result.
}
\description{
Expect that a CSS selector resolves to at least one visible element.
}
\section{Functions}{
\itemize{
\item \code{app_driver_expect_hidden()}: Check if a selector is hidden for a given timeout.

}}
\keyword{internal}
