% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{are_all_picks_fixed}
\alias{are_all_picks_fixed}
\title{Check if picks object has all elements fixed}
\usage{
are_all_picks_fixed(picks)
}
\arguments{
\item{picks}{object of class (\code{picks})}
}
\description{
Check if picks object has all elements fixed
}
\keyword{internal}
