% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{check_level}
\alias{check_level}
\alias{assert_level}
\title{Assert level}
\usage{
check_level(x, class)

assert_level(x, class, .var.name = checkmate::vname(x), add = NULL)
}
\arguments{
\item{x}{\code{picks} object}

\item{class}{Class of the last element of picks}
}
\value{
For \code{check_level} a logical value or a string.
For \code{assert_level} invisibly the object checked or an error.
}
\description{
Assert level
}
\examples{
x <- picks(datasets(), variables(), values())
assert_level(x, "values")
}
