% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{check_picks}
\alias{check_picks}
\alias{assert_picks}
\title{Assert picks contains a specific element}
\usage{
check_picks(x, datasets = TRUE, variables = FALSE, values = FALSE)

assert_picks(
  x,
  datasets = TRUE,
  variables = FALSE,
  values = FALSE,
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{\code{picks} object}

\item{datasets}{(\code{logical(1)}) whether to check for the presence of a \code{datasets} element.}

\item{variables}{(\code{logical(1)}) whether to check for the presence of a \code{variables} element}

\item{values}{(\code{logical(1)}) whether to check for the presence of a \code{values} element}

\item{.var.name}{[\code{character(1)}]\cr
The custom name for \code{x} as passed to any \code{assert*} function.
Defaults to a heuristic name lookup.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
For \code{check_picks} a logical value or a string.
For \code{assert_picks} invisibly the object checked or an error.
}
\description{
Check whether a \code{picks} object contains at least one element of the given
class (\code{"datasets"}, \code{"variables"}, or \code{"values"}).
}
\examples{
p <- picks(datasets(), variables(), values())
assert_picks(p, datasets = TRUE)
assert_picks(p, variables = TRUE)
assert_picks(p, values = TRUE)

p <- picks(datasets(), variables())
check_picks(p, values = TRUE)
}
