% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{check_picks_has}
\alias{check_picks_has}
\alias{assert_picks_has}
\title{Assert picks contains a specific element}
\usage{
check_picks_has(x, element = c("datasets", "variables", "values"))

assert_picks_has(
  x,
  element = c("datasets", "variables", "values"),
  .var.name = checkmate::vname(x),
  add = NULL
)
}
\arguments{
\item{x}{\code{picks} object}

\item{element}{(\code{character(1)}) one of \code{"datasets"}, \code{"variables"}, or
\code{"values"}.}

\item{.var.name}{[\code{character(1)}]\cr
The custom name for \code{x} as passed to any \code{assert*} function.
Defaults to a heuristic name lookup.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
For \code{check_picks_has} a logical value or a string.
For \code{assert_picks_has} invisibly the object checked or an error.
}
\description{
Check whether a \code{picks} object contains at least one element of the given
class (\code{"datasets"}, \code{"variables"}, or \code{"values"}).
}
\examples{
p <- picks(datasets(), variables(), values())
assert_picks_has(p, "datasets")
assert_picks_has(p, "variables")
assert_picks_has(p, "values")
}
