% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_containers.R
\name{badge_fixed}
\alias{badge_fixed}
\title{Fixed badge}
\usage{
badge_fixed(id, label, content)
}
\arguments{
\item{id}{(\code{character(1)}) shiny module's id}

\item{label}{(\code{shiny.tag}) Label displayed on a badge.}

\item{content}{(\code{shiny.tag}) Content of a drop-down.}
}
\value{
A \code{shiny.tag} object representing a fixed badge component to display the selected option.
}
\description{
Fixed badge
}
\keyword{internal}
