% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_picks_helper}
\alias{create_picks_helper}
\title{Creation of picks object that does not override a dataset if already exists}
\usage{
create_picks_helper(datasets = NULL, x)
}
\arguments{
\item{datasets}{(\code{\link[=datasets]{datasets()}} object) to use if \code{x} does not already have a dataset.}

\item{x}{(\code{pick} or \code{picks} object) to ensure has a dataset.}
}
\value{
a \code{picks} object with a dataset, either from \code{x} or from \code{datasets}.
}
\description{
Utility function for applying a user-input for variables to the data selected.
}
\examples{
create_picks_helper("ADTTE", x = picks(datasets("ADSL", "ADSL"), variables("SEX")))
create_picks_helper(datasets("ADSL", "ADSL"), x = variables("SEX", "SEX"))
}
