% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-call_utils.R
\name{.call_dplyr_filter}
\alias{.call_dplyr_filter}
\title{\code{dplyr::filter} call}
\usage{
.call_dplyr_filter(x)
}
\arguments{
\item{x}{(\code{list}) containing \code{variables} and \code{values}}
}
\description{
Create a \code{dplyr::filter} call
}
\keyword{internal}
