% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-call_utils.R
\name{.call_dplyr_select}
\alias{.call_dplyr_select}
\title{\code{dplyr::select} call}
\usage{
.call_dplyr_select(dataname, variables)
}
\arguments{
\item{dataname}{(\code{character(1)}) name of the dataset}

\item{variables}{(\code{list} of \code{character}) variables to select. If list is named then
variables will be renamed if their name is different than its value
(this produces a call \verb{select(..., <name> = <value>)}).}
}
\description{
Create \code{dplyr::select} call from \code{dataname} and \code{variables}
}
