% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-picks.R
\name{.pick}
\alias{.pick}
\title{Pick class constructor}
\usage{
.pick(
  choices,
  selected,
  multiple = length(selected) > 1,
  ordered = FALSE,
  fixed = FALSE,
  ...
)
}
\arguments{
\item{choices}{(\code{tidyselect::language} or \code{character})
Available values to choose.}

\item{selected}{(\code{tidyselect::language} or \code{character})
Choices to be selected.}

\item{multiple}{(\code{logical(1)}) if more than one selection is possible.}

\item{ordered}{(\code{logical(1)}) if the selected should follow the selection order. If \code{FALSE}
\code{selected} returned from \code{srv_module_input()} would be ordered according to order in \code{choices}.}

\item{fixed}{(\code{logical(1)}) selection will be fixed and not possible to change interactively.}

\item{...}{for \code{picks(...)}: hierarchical structure that contains \code{datasets()} as first element and optionally \code{variables()} and \code{values()}

for \code{variables(...)} and \code{values(...)}: additional arguments delivered to \code{pickerInput}}
}
\description{
Create a \code{pick} object
}
\keyword{internal}
