% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_picks.R
\name{.update_rv}
\alias{.update_rv}
\title{Update reactive values with log}
\usage{
.update_rv(rv, value, log)
}
\arguments{
\item{rv}{(\code{reactiveVal})}

\item{value}{(\code{vector})}

\item{log}{(\code{character(1)}) message to \code{log_debug}}
}
\value{
updated \code{reactiveVal} with new value if it differs from the old one, otherwise unchanged
}
\description{
Update reactive values only if values differ to avoid unnecessary reactive trigger
}
\keyword{internal}
