% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-ui_containers.R
\name{fixed_picks_container}
\alias{fixed_picks_container}
\title{Create ui component for fixed ui picks without user selection}
\usage{
fixed_picks_container(id, label)
}
\arguments{
\item{id}{(\code{character(1)}) shiny module's id}

\item{label}{(\code{shiny.tag}) Label displayed on the component}

\item{class}{(\code{character(1)}) string with the css class of the component}
}
\description{
Create ui component for fixed ui picks without user selection
}
\keyword{internal}
