% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_functions_pick}
\alias{helper_functions_pick}
\alias{is_pick_multiple}
\alias{is_pick_fixed}
\alias{is_pick_ordered}
\title{Helper functions for pick}
\usage{
is_pick_multiple(x)

is_pick_fixed(x)

is_pick_ordered(x)
}
\arguments{
\item{x}{(\code{datasets}, \code{variables} or \code{values}) pick to check.}
}
\value{
\code{TRUE} if the pick has the attribute set to \code{TRUE},
\code{FALSE} otherwise.
}
\description{
Helper functions for pick objects generated from
\code{\link[=datasets]{datasets()}}, \code{\link[=variables]{variables()}} or \code{\link[=values]{values()}}:

\itemize{
\item \code{is_pick_multiple()} checks if a pick has the \code{multiple} attribute set to \code{TRUE}.
}

\itemize{
\item \code{is_pick_fixed()} checks if a pick has the \code{fixed} attribute set to \code{TRUE}.
}

\itemize{
\item \code{is_pick_ordered()} checks if a pick has the \code{ordered} attribute set to \code{TRUE}.
}
}
\examples{
p <- picks(datasets("iris"), variables(), values())

is_pick_multiple(p$variables)

is_pick_fixed(p$variables)

is_pick_ordered(p$variables)
}
