% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pick_single_variable}
\alias{pick_single_variable}
\title{Checks picks object}
\usage{
pick_single_variable(picks)
}
\arguments{
\item{picks}{A \code{picks} object}
}
\value{
The picks object with only being able to select one variable.
}
\description{
Enforce a single variable selection on a picks object.
}
\examples{
p <- picks(datasets(), variables(multiple = TRUE))
pick_single_variable(p)
}
