% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{picks_datanames}
\alias{picks_datanames}
\title{Extract datanames from list of picks}
\usage{
picks_datanames(...)
}
\arguments{
\item{...}{One or more picks object.}
}
\value{
The names of the datasets used
}
\description{
Helper to decide which datanames are needed for a teal module from the user input.
}
\examples{
picks_datanames(
  picks(
    datasets("ADSL", "ADSL"),
    variables("SEX")
  ),
  picks(datasets("ADTTE", "ADTTE"))
)
}
