% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-tidyselect-helpers.R
\name{ranged}
\alias{ranged}
\alias{could_be_range}
\title{Select a range}
\usage{
ranged(min, max)

could_be_range(x)
}
\arguments{
\item{min}{Minimal value.}

\item{max}{Maximal value.}

\item{x}{Some data}
}
\description{
Helper functions for working with ranges. Main function is \code{ranged}.
}
\section{Functions}{
\itemize{
\item \code{could_be_range()}: Check if some values could be a ranged output.

}}
\examples{
p <- picks(
  datasets(is.data.frame, is.data.frame),
  variables(is.numeric, 1),
  values(where(~ .x > 5), ranged(20, NA))
)
could_be_range(mtcars$disp)
could_be_range(iris$Species)
is_range(p$values$selected)
r <- resolver(data = list("mtcars label" = mtcars), x = p)
is_range(r$values$selected)
}
