% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.picks-package.R
\docType{package}
\name{teal.picks-package}
\alias{teal.picks}
\alias{teal.picks-package}
\title{teal.picks: Dataset and Variable Picker and Merge Module for 'teal' Applications}
\description{
Allows users to interactively select datasets, variables, and values within 'teal' applications using a 'tidyselect'-style interface. Selected picks can be merged and transformed into analysis-ready data within 'teal' modules.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmaverse/teal.picks/}
  \item \url{https://pharmaverse.github.io/teal.picks/}
  \item Report bugs at \url{https://github.com/pharmaverse/teal.picks/issues}
}

}
\author{
\strong{Maintainer}: Dony Unardi \email{unardid@gene.com} [reviewer]

Authors:
\itemize{
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com} (\href{https://orcid.org/0000-0001-9533-457X}{ORCID})
  \item Andre Verissimo \email{andre.verissimo@roche.com} (\href{https://orcid.org/0000-0002-2212-339X}{ORCID})
  \item Marcin Kosinski \email{marcin.kosinski.mk1@roche.com}
  \item Lluís Revilla Sancho \email{lluis.revilla_sancho@roche.com} (\href{https://orcid.org/0000-0001-9747-2570}{ORCID})
  \item Oriol Senan \email{oriol.senan@external.roche.com} (\href{https://orcid.org/0000-0002-9621-3371}{ORCID})
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
