% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-shinytest2.R
\name{app_driver_teal_picks_exports}
\alias{app_driver_teal_picks_exports}
\title{Read all teal.picks exported values for a module namespace.}
\usage{
app_driver_teal_picks_exports(app_driver, pick_id)
}
\arguments{
\item{app_driver}{App driver object.}

\item{pick_id}{\code{character(1)} teal.picks id.}
}
\value{
Named list of module-scoped exported values.
}
\description{
The module namespace is inferred from the summary badge id in the \code{DOM},
then used to filter exported values.
}
\keyword{internal}
