% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ensure_picks_datasets}
\alias{ensure_picks_datasets}
\title{Creation of picks object that does not override a dataset if already exists}
\usage{
ensure_picks_datasets(datasets = NULL, x, ...)
}
\arguments{
\item{datasets}{(\code{\link[=datasets]{datasets()}} object) to use if \code{x} does not already have a dataset.}

\item{x}{(\code{pick} or \code{picks} object) to ensure has a dataset.}

\item{...}{(\code{pick} or \code{picks} object) that will be appended to the \code{datasets}.}
}
\value{
a \code{picks} object with a dataset, either from \code{x} or from \code{datasets} and other information added.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Utility function for applying a user-input for variables to the data selected.
}
\examples{
ensure_picks_datasets("ADTTE", x = picks(datasets("ADSL", "ADSL"), variables("SEX")))
ensure_picks_datasets(datasets("ADSL", "ADSL"), x = variables("SEX", "SEX"))
ensure_picks_datasets(datasets("ADSL", "ADSL"), x = variables("SEX", "SEX"), values(c("F", "M"), "F"))
}
