% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-tidyselect-helpers.R
\name{ranged}
\alias{ranged}
\title{Select a range}
\usage{
ranged(min, max)
}
\arguments{
\item{min}{Minimal value.}

\item{max}{Maximal value.}
}
\description{
Helper functions for working with ranges. Main function is \code{ranged}.
}
\examples{
p <- picks(
  datasets(is.data.frame, is.data.frame),
  variables(is.numeric, 1),
  values(tidyselect::where(~ .x > 5), ranged(20, NA))
)
resolver(data = list("mtcars label" = mtcars), x = p)
}
